+++ apps/docs/components/index.tsx
@@ -1,6 +1,1 @@
-export * from '@quantit/qui-react/src/components';
-export {
-  Button as ButtonNew,
-  IconButton as IconButtonNew,
-  Icon as IconNew,
-} from 'qui-base/lib/main';
+export * from 'qui-base/lib/main';