+++ packages/qui-core/src/index.tsx
@@ -1,4 +1,0 @@
-import * as React from 'react';
-export * from './components';
-export * from './styles/theme';
-export { QuiThemeProvider } from './styles/QuiThemeProvider';