+++ packages/qui-core/src/styles/theme/index.ts
@@ -1,5 +1,0 @@
-export * from './const';
-export type { Color, ColorRange } from './color.type';
-export type { ThemeMode, PaletteOptions } from './palette.type';
-export type { TypographyVariantType } from './typography.type';
-export type { QuiTheme } from './theme.type';