+++ packages/qui-core/src/styles/tokens/divider/divider.token.ts
@@ -1,6 +1,0 @@
-import { DefaultTheme } from 'styled-components';
-
-export const dividerToken = (theme: DefaultTheme) => ({
-  darken: theme.color.semantic.border_low,
-  lighten: theme.color.semantic.border_medium,
-});