+++ packages/qui-core/src/styles/tokens/scrollbar/index.ts
@@ -1,2 +1,0 @@
-export * from './scrollbar.token';
-export * from './scrollbar.token.util';