+++ packages/qui-core/src/styles/tokens/scrollbar/scrollbar.token.ts
@@ -1,5 +1,0 @@
-import { DefaultTheme } from 'styled-components';
-
-export const scrollbarToken = (theme: DefaultTheme) => ({
-  container: theme.color.black_alpha[70],
-});