+++ packages/qui-core/src/styles/tokens/state-overlay/index.ts
@@ -1,2 +1,0 @@
-export * from './state-overlay.token';
-export * from './state-overlay.token.util';