+++ packages/qui-core/src/styles/tokens/switch/index.ts
@@ -1,7 +1,0 @@
-export * from './checkbox.token';
-export * from './checkbox.token.util';
-export * from './common.token';
-export * from './tab.token';
-export * from './tab.token.util';
-export * from './toggle.token';
-export * from './toggle.token.util';