+++ packages/qui-core/src/styles/utils/isDarkMode.util.ts
@@ -1,6 +1,0 @@
-export function isDarkMode(): boolean {
-  if (window.matchMedia?.('(prefers-color-scheme: dark)').matches) {
-    return true;
-  }
-  return false;
-}